+++ src/resources/AuditEvents.ts
@@ -1,0 +1,6 @@
+import TelnyxResource from '../TelnyxResource';
+
+export const AuditEvents = TelnyxResource.extend({
+  path: 'audit_events',
+  includeBasic: ['list'],
+});