+++ src/resources/Events.ts
@@ -1,6 +1,0 @@
-import TelnyxResource from '../TelnyxResource';
-
-export const Events = TelnyxResource.extend({
-  path: 'events',
-  includeBasic: ['list', 'retrieve'],
-});