+++ src/resources/PhoneNumbersCsvDownloads.ts
@@ -1,0 +1,6 @@
+import TelnyxResource from '../TelnyxResource';
+
+export const PhoneNumbersCsvDownloads = TelnyxResource.extend({
+  path: '/phone_numbers/csv_downloads',
+  includeBasic: ['list', 'create', 'retrieve'],
+});