+++ src/test/resources/DetailRecords.test.ts
@@ -4,7 +4,7 @@ const telnyx = testUtils.getTelnyxMock();
 
 describe('DetailRecords Resource', function () {
   function responseFn(response: ResponsePayloadList) {
-    expect(response.data).toHaveProperty('data');
+    expect(response).toHaveProperty('data');
     for (let index = 0; index < response.data.length; index++) {
       const row = response.data[index];
       expect(row).toHaveProperty('record_type');
@@ -12,11 +12,14 @@ describe('DetailRecords Resource', function () {
     }
   }
 
-  describe.skip('query', function () {
+  describe('search', function () {
     test('Sends the correct request', function () {
       // @ts-expect-error TODO: import .d.ts files under src/test folder
       return telnyx.detailRecords
-        .query({filter: {record_type: 'messaging'}})
+        .search({
+          // limitation on prism mock
+          'filter%5brecord_type%5d': 'messaging',
+        })
         .then(responseFn);
     });
   });